cloud-terraform/face_cut/requirements.txt
requests==2.31.0
boto3==1.29.6
pillow==10.1.0
ydb==3.5.2